refactor(tests): move ping parsing tests to test_pinger.py
